Two LFO waveforms for great happiness (#3)
CV2 input now determines shape of LFO waveform as well as frequency.
Counterclockwise (0-2.5v) is a triangle wave as before, Clockwise (2.5v to
5v) is a sawtooth wave.

There is no clever alignment of LFO phase so the sawtooth wave is going to
be higher frequency than the triangle. Deal with it.

Given the gate output on LFO rising makes a square wave LFO we now have
three LFO waveforms - not bad.